v1.0 Android and iOS update

+++ iOSSDKModule/Classes/ComAlfrescoAppceleratorSdkListingContextProxy.m
@@ -43,7 +43,7 @@
 
     NSNumber* maxItems = arg;
     
-    self.listingContext = [[AlfrescoListingContext alloc]initWithMaxItems:maxItems];
+    self.listingContext = [[AlfrescoListingContext alloc]initWithMaxItems:maxItems.intValue];
 }
 
 
@@ -52,7 +52,7 @@
     NSNumber* maxItems = [args objectAtIndex:0];
     NSNumber* skipCount = [args objectAtIndex:1];
     
-    self.listingContext = [[AlfrescoListingContext alloc]initWithMaxItems:maxItems skipCount:skipCount];
+    self.listingContext = [[AlfrescoListingContext alloc]initWithMaxItems:maxItems.intValue skipCount:skipCount.intValue];
 }
 
 
@@ -72,7 +72,7 @@
     NSString* sortProperty = [args objectAtIndex:2];
     BOOL sortAscending = ([args objectAtIndex:3] > 0);
     
-    self.listingContext = [[AlfrescoListingContext alloc]initWithMaxItems:maxItems skipCount:skipCount sortProperty:sortProperty
+    self.listingContext = [[AlfrescoListingContext alloc]initWithMaxItems:maxItems.intValue skipCount:skipCount.intValue sortProperty:sortProperty
                                                                      sortAscending:sortAscending];
 }
